requirements.txt
Flask
numpy
pillow
cryptography
boto3
python-dotenv
